Reorganized project; added scss modules support
