+++ auction/src/styles/App.module.scss
@@ -1,0 +1,5 @@
+.class {
+    background: #333;
+    color: white;
+    padding: 8px;
+}
